test: 불필요한 에너테이션 옵션 제거 및 설정 분리 TODAK-209
- Interceptor.class 설정 제거
- test security config 분리
