+++ src/test/java/com/heartsave/todaktodak_api/ai/webhook/test_config/TestInterceptorSecurityConfig.java
@@ -1,0 +1,27 @@
+package com.heartsave.todaktodak_api.ai.webhook.test_config;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.web.SecurityFilterChain;
+
+@TestConfiguration
+public class TestInterceptorSecurityConfig {
+  @Bean
+  public WebSecurityCustomizer webSecurityCustomizer() {
+    return (web) -> web.ignoring().requestMatchers("/api/v1/webhook/ai/**");
+  }
+
+  @Bean
+  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+    return http.csrf(csrf -> csrf.disable())
+        .authorizeHttpRequests(
+            auth ->
+                auth.requestMatchers("/api/v1/webhook/ai/**")
+                    .permitAll()
+                    .anyRequest()
+                    .authenticated())
+        .build();
+  }
+}
